backend/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0
python-jose==3.3.0
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
requests==2.32.3
beautifulsoup4==4.12.3
pypdf==4.3.1
openai==1.40.2
httpx==0.27.2
numpy==2.0.2